+++ Homework_2/RayTracer.cpp
@@ -13,7 +13,7 @@ auto RayTracer::castRaysSingleThreaded(const sf::Vector2f &lightPos, int numRays
     results.resize(numRays);
     for (int i = 0; i < numRays; ++i)
     {
-        float angle = (2.0f * M_PI * i) / numRays;
+        float angle = static_cast<float>((2.0f * M_PI * i) / numRays);
         Ray ray;
         ray.origin = lightPos;
         ray.direction = {std::cos(angle), std::sin(angle)};
@@ -29,7 +29,7 @@ auto RayTracer::castRaysOpenMP(const sf::Vector2f &lightPos, int numRays, const 
 #pragma omp parallel for schedule(static)
     for (int i = 0; i < numRays; ++i)
     {
-        float angle = (2.0f * M_PI * i) / numRays;
+        float angle = static_cast<float>((2.0f * M_PI * i) / numRays);
         Ray ray;
         ray.origin = lightPos;
         ray.direction = {std::cos(angle), std::sin(angle)};
